upload color
